+++ ObjCTests/Tests/TemperatureTests.m
@@ -9,7 +9,7 @@
 #import <Unitility/Temperature.h>
 #import <XCTest/XCTest.h>
 
-static CGFloat const kAccuracy = 0.01f;
+static float const kAccuracy = 0.01f;
 
 @interface TemperatureTests : XCTestCase
